+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Version - 2]
+## Feature [2.11.0] - 2024-03-07
+- add new feature approve special request 
 
 ## Feature [2.10.1] - 2024-01-15
 - upload file